requirements.txt
python-dotenv
pandas 
numpy
matplotlib
scikit-learn
seaborn
pymongo
certifi
pymongo[srv]==3.6
dill
pyaml

-e .

# '''
# this -e . file refers to the setup.py file when we install the libraries and 
# help to setup the project 
# '''
